requirements.txt
tensorflow>=2.0.0
flask>=2.0.0
Pillow>=9.0.0
numpy>=1.20.0
werkzeug>=2.0.0S